metode getPotencia cambiat a double

+++ src/main/java/org/JavaCar/Motor.java
@@ -17,7 +17,7 @@ public class Motor {
         this.tipus = tipus;
     }
 
-    public int getPotencia() {
+    public double getPotencia() {
         return potencia;
     }
